+++ src/main/java/com/caffeine/gwanghwamun/domain/menu/controller/MenuOptionController.java
@@ -1,14 +1,14 @@
-package com.caffeine.gwanghwamun.domain.menuoption.controller;
+package com.caffeine.gwanghwamun.domain.menu.controller;
 
 import com.caffeine.gwanghwamun.common.response.ApiResponse;
 import com.caffeine.gwanghwamun.common.response.ResponseUtil;
 import com.caffeine.gwanghwamun.common.success.SuccessCode;
-import com.caffeine.gwanghwamun.domain.menuoption.dto.request.MenuOptionCreateReqDTO;
-import com.caffeine.gwanghwamun.domain.menuoption.dto.request.MenuOptionSoldOutReqDTO;
-import com.caffeine.gwanghwamun.domain.menuoption.dto.request.MenuOptionUpdateReqDTO;
-import com.caffeine.gwanghwamun.domain.menuoption.dto.request.MenuOptionVisibilityReqDTO;
-import com.caffeine.gwanghwamun.domain.menuoption.dto.response.MenuOptionResDTO;
-import com.caffeine.gwanghwamun.domain.menuoption.service.MenuOptionService;
+import com.caffeine.gwanghwamun.domain.menu.dto.request.MenuOptionCreateReqDTO;
+import com.caffeine.gwanghwamun.domain.menu.dto.request.MenuOptionSoldOutReqDTO;
+import com.caffeine.gwanghwamun.domain.menu.dto.request.MenuOptionUpdateReqDTO;
+import com.caffeine.gwanghwamun.domain.menu.dto.request.MenuOptionVisibilityReqDTO;
+import com.caffeine.gwanghwamun.domain.menu.dto.response.MenuOptionResDTO;
+import com.caffeine.gwanghwamun.domain.menu.service.MenuOptionService;
 import com.caffeine.gwanghwamun.domain.user.entity.User;
 import io.swagger.v3.oas.annotations.Operation;
 import java.util.UUID;
