+++ src/main/java/com/caffeine/gwanghwamun/domain/menu/dto/request/MenuOptionUpdateReqDTO.java
@@ -1,4 +1,4 @@
-package com.caffeine.gwanghwamun.domain.menuoption.dto.request;
+package com.caffeine.gwanghwamun.domain.menu.dto.request;
 
 import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
