+++ src/main/java/com/caffeine/gwanghwamun/domain/menu/dto/request/MenuOptionVisibilityReqDTO.java
@@ -1,0 +1,3 @@
+package com.caffeine.gwanghwamun.domain.menu.dto.request;
+
+public record MenuOptionVisibilityReqDTO(Boolean hidden) {}
