+++ src/main/java/com/caffeine/gwanghwamun/domain/menu/repository/MenuOptionRepository.java
@@ -1,8 +1,9 @@
-package com.caffeine.gwanghwamun.domain.menuoption.repository;
+package com.caffeine.gwanghwamun.domain.menu.repository;
 
-import com.caffeine.gwanghwamun.domain.menuoption.entity.MenuOption;
+import com.caffeine.gwanghwamun.domain.menu.entity.MenuOption;
 import java.util.Optional;
 import java.util.UUID;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
