--- src/main/java/com/caffeine/gwanghwamun/domain/menuoption/dto/request/MenuOptionVisibilityReqDTO.java
@@ -1,3 +1,0 @@
-package com.caffeine.gwanghwamun.domain.menuoption.dto.request;
-
-public record MenuOptionVisibilityReqDTO(Boolean hidden) {}
